pyproject.toml
[project]
name = "rag-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.3.5",
    "fastapi>=0.122.0",
    "jinja2>=3.1.6",
    "langchain>=0.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
